Update scoring system to show winner of the game
The playerRound function returns the integer -1 to represent the
player losing a round, and the integer 1 to represent the player
winning a round.

The game function tallies the overall score. If the score is
negative, the computer wins, and if the score is positive, the
player wins. If the score is zero, both the player and computer
tied.