Refactor scoring system
An array containing the player and computer scores is used in the
scoring system. The scoring system function checks if either the
player or the computer won, and tallies the player and computer
scores.

The function also stops the score from increasing if either the
player or the computer reached a score of 5.

The function also shows the current scores of both the player and
the computer.