Implement game end check and reset button
The end game function checks if either the player or the computer
won the game. If either won, it removes the roshambo buttons and
replaces them with the reset button.

The reset button resets the scores to zero and updates the result
texts through the reset game function. The reset game function
removes the reset button and adds the roshambo buttons.